refactor: Rename @n8n/utils files to kebab-case and enforce with ESLint
