+++ packages/@n8n/utils/src/index.ts
@@ -1,17 +1,17 @@
 export * from './assert';
 export * from './event-bus';
-export * from './event-queue';
+export * from './create-event-queue';
 export * from './retry';
-export * from './workflowId';
-export * from './number/smartDecimal';
-export * from './search/reRankSearchResults';
-export * from './search/sublimeSearch';
-export * from './sort/sortByProperty';
+export * from './generate-nano-id';
+export * from './number/smart-decimal';
+export * from './search/re-rank-search-results';
+export * from './search/sublime-search';
+export * from './sort/sort-by-property';
 export * from './string/truncate';
-export * from './files/sanitize';
-export * from './files/path';
+export * from './files/sanitize-filename';
+export * from './files/is-windows-file-path';
 export * from './placeholder';
-export * from './jwt';
+export * from './get-jwt-expiry';
 export * from './scrub-secrets';
 export type * from './types';
-export * from './validation';
+export * from './is-record';
