+++ packages/@n8n/utils/src/is-record.ts
