+++ packages/@n8n/utils/src/number/smart-decimal.ts
