+++ packages/frontend/editor-ui/src/features/execution/insights/components/InsightsSummary.vue
@@ -10,7 +10,7 @@ import type { DateValue } from '@internationalized/date';
 import type { InsightsSummary } from '@n8n/api-types';
 import { N8nIcon, N8nTooltip } from '@n8n/design-system';
 import { useI18n } from '@n8n/i18n';
-import { smartDecimal } from '@n8n/utils/number/smartDecimal';
+import { smartDecimal } from '@n8n/utils/number/smart-decimal';
 import { computed, useCssModule } from 'vue';
 import { I18nT } from 'vue-i18n';
 import { useRoute } from 'vue-router';
