+++ packages/frontend/editor-ui/src/features/execution/insights/components/charts/InsightsChartFailed.vue
@@ -2,7 +2,7 @@
 import { useI18n } from '@n8n/i18n';
 import { generateBarChartOptions } from '@/features/execution/insights/chartjs.utils';
 import { GRANULARITY_DATE_FORMAT_MASK } from '@/features/execution/insights/insights.constants';
-import { smartDecimal } from '@n8n/utils/number/smartDecimal';
+import { smartDecimal } from '@n8n/utils/number/smart-decimal';
 import { useCssVar } from '@vueuse/core';
 import type { ChartData } from 'chart.js';
 import { computed } from 'vue';
