+++ packages/frontend/editor-ui/src/features/execution/insights/components/tables/InsightsTableWorkflows.vue
@@ -8,7 +8,7 @@ import {
 } from '@/features/execution/insights/insights.utils';
 import type { InsightsByWorkflow } from '@n8n/api-types';
 import type { TableHeader } from '@n8n/design-system/components/N8nDataTableServer';
-import { smartDecimal } from '@n8n/utils/number/smartDecimal';
+import { smartDecimal } from '@n8n/utils/number/smart-decimal';
 import { useTelemetry } from '@/app/composables/useTelemetry';
 import { VIEWS } from '@/app/constants';
 import { computed, defineAsyncComponent, ref, watch } from 'vue';
